+++ github-workflow-commands.cabal
@@ -31,16 +31,16 @@ library
   exposed-modules:
       GitHub.Workflow.Command.Annotation
       GitHub.Workflow.Command.Annotation.Column
-      GitHub.Workflow.Command.Annotation.Debug
-      GitHub.Workflow.Command.Annotation.Error
+      GitHub.Workflow.Command.Annotation.Commands.Debug
+      GitHub.Workflow.Command.Annotation.Commands.Error
+      GitHub.Workflow.Command.Annotation.Commands.Generic
+      GitHub.Workflow.Command.Annotation.Commands.Notice
+      GitHub.Workflow.Command.Annotation.Commands.Warning
       GitHub.Workflow.Command.Annotation.File
-      GitHub.Workflow.Command.Annotation.GenericAnnotation
       GitHub.Workflow.Command.Annotation.Line
       GitHub.Workflow.Command.Annotation.Location
-      GitHub.Workflow.Command.Annotation.Notice
       GitHub.Workflow.Command.Annotation.Position
       GitHub.Workflow.Command.Annotation.Properties
-      GitHub.Workflow.Command.Annotation.Warning
       GitHub.Workflow.Command.Isomorphism.Natural
       GitHub.Workflow.Command.Isomorphism.Text
       GitHub.Workflow.Command.Syntax
